+++ server/impl/src/main/java/com/walmartlabs/concord/server/A.java
@@ -1,0 +1,4 @@
+package com.walmartlabs.concord.server;
+
+public class A {
+}