+++ server/impl/src/main/java/com/walmartlabs/concord/server/Utils.java
@@ -20,6 +20,8 @@ package com.walmartlabs.concord.server;
  * =====
  */
 
+import java.util.Properties;
+
 public final class Utils {
 
     public static String[] toString(Enum<?>... e) {
@@ -28,6 +30,30 @@ public final class Utils {
             as[i] = e[i].toString();
         }
         return as;
+    }
+
+    public static String getEnv(String key, String defaultValue) {
+        String s = System.getenv(key);
+        if (s == null) {
+            return defaultValue;
+        }
+        return s;
+    }
+
+    public static long getLong(Properties props, String key, long defaultValue) {
+        String s = props.getProperty(key);
+        if (s == null) {
+            return defaultValue;
+        }
+        return Long.parseLong(s);
+    }
+
+    public static boolean getBoolean(Properties props, String key, boolean defaultValue) {
+        String s = props.getProperty(key);
+        if (s == null) {
+            return defaultValue;
+        }
+        return Boolean.parseBoolean(s);
     }
 
     private Utils() {